Run the main-publication invariant tests on FwData too
The single-main guards are enforced in the backend-agnostic validation wrapper, so move them to PublicationsTestsBase. FwData maps IsMain from the read-only IsProtected, so the FwData test seeds a main at the LCM layer. The promotion and merge-convergence tests stay CRDT-only.
