+++ backend/FwLite/LcmCrdt.Tests/MiniLcmTests/PublicationsTests.cs
@@ -21,37 +21,6 @@ public class PublicationsTests : PublicationsTestsBase
     }
 
     [Fact]
-    public async Task UpdatePublication_CannotTurnOffIsMain()
-    {
-        var main = await Api.CreatePublication(new Publication { Id = Guid.NewGuid(), Name = { { "en", "Main" } }, IsMain = true });
-
-        var act = () => Api.UpdatePublication(main.Id, new UpdateObjectInput<Publication>().Set(p => p.IsMain, false));
-
-        await act.Should().ThrowAsync<FluentValidation.ValidationException>();
-    }
-
-    [Fact]
-    public async Task UpdatePublication_CannotPromoteSecondMain()
-    {
-        await Api.CreatePublication(new Publication { Id = Guid.NewGuid(), Name = { { "en", "Main" } }, IsMain = true });
-        var other = await Api.CreatePublication(new Publication { Id = Guid.NewGuid(), Name = { { "en", "Pocket" } } });
-
-        var act = () => Api.UpdatePublication(other.Id, new UpdateObjectInput<Publication>().Set(p => p.IsMain, true));
-
-        await act.Should().ThrowAsync<InvalidOperationException>();
-    }
-
-    [Fact]
-    public async Task CreatePublication_CannotCreateSecondMain()
-    {
-        await Api.CreatePublication(new Publication { Id = Guid.NewGuid(), Name = { { "en", "Main" } }, IsMain = true });
-
-        var act = () => Api.CreatePublication(new Publication { Id = Guid.NewGuid(), Name = { { "en", "Second" } }, IsMain = true });
-
-        await act.Should().ThrowAsync<InvalidOperationException>();
-    }
-
-    [Fact]
     public async Task PublicationSync_PromotesExistingPublicationToMain()
     {
         var pub = await Api.CreatePublication(new Publication { Id = Guid.NewGuid(), Name = { { "en", "Pocket" } } });
